BBC Radio streams updated
They've changed here:
https://www.bbc.co.uk/blogs/internet/entries/481e7233-0aea-4b15-8ace-878ce549108c

There is a link to:
https://gist.github.com/bpsib/67089b959e4fa898af69fea59ad74bc3

Because radio-pi uses mplayer I've done: curl of the pls.pls files, then
get the File1 and put it into stations.json for mplayer to play.

cvlc would be a better option but it needs code changes and I don't see
how to get the song name (e.g. for Smooth FM) as easy as using mplayer
(parsing the standard output). To be done another time